move summarizeItems to separate file
